Udpated ip.sh
Turns wifi off if ethernet is connected. Turns wifi back on and updates caffeinate plugin if "Save and caffeinate" is selected. Differentiates between the wifi being off or something else.  Writes to a file on desktop every times it updates (used to see how frequently it updates when no interval is specified)